:sparkles: Use openapi 3 format in schema endpoint and not swagger 2
